Fix conversion error
This causes a crash of the provider

+++ builtin/providers/cloudstack/provider.go
@@ -58,7 +58,7 @@ func providerConfigure(d *schema.ResourceData) (interface{}, error) {
 		ApiURL:    d.Get("api_url").(string),
 		ApiKey:    d.Get("api_key").(string),
 		SecretKey: d.Get("secret_key").(string),
-		Timeout:   d.Get("timeout").(int64),
+		Timeout:   int64(d.Get("timeout").(int)),
 	}
 
 	return config.NewClient()
